+++ src/Raw_Material_Department/Raw_Material.hpp
@@ -22,7 +22,8 @@ public:
 	static int getMaterialCount();
 
 	// Manipulate Materials
-	void addMaterial(string materialType, string materialQulaity, float cost, int suppliedQuantity, int safetystock, int availableStock, int orderedStock, int stockDuration);
+	void addMaterial(string materialType, string materialQulaity, long cost, int suppliedQuantity, int safetystock,
+					 int availableStock, int orderedStock, int stockDuration);
 	void removeMaterial(int materialID);
 	void getMaterial(int materialID);
